Add more jar inside faces

+++ src/main/kotlin/com/kneelawk/jarit/client/util/RenderExtensions.kt
@@ -6,7 +6,8 @@ import net.minecraft.client.util.SpriteIdentifier
 import net.minecraft.util.math.Matrix4f
 
 class SpriteRenderScope(private val sprite: Sprite) {
-    fun VertexConsumer.frameUV(u: Double, v: Double): VertexConsumer = uv(sprite.getFrameU(u), sprite.getFrameV(v))
+    fun VertexConsumer.frameUV(u: Int, v: Int): VertexConsumer =
+        uv(sprite.getFrameU(u.toDouble()), sprite.getFrameV(v.toDouble()))
 }
 
 inline fun <R> Sprite.scope(f: SpriteRenderScope.() -> R): R {
@@ -17,5 +18,5 @@ inline fun <R> SpriteIdentifier.scope(f: SpriteRenderScope.() -> R): R {
     return sprite.scope(f)
 }
 
-fun VertexConsumer.pixelVertex(matrix: Matrix4f, x: Double, y: Double, z: Double): VertexConsumer =
+fun VertexConsumer.pixelVertex(matrix: Matrix4f, x: Int, y: Int, z: Int): VertexConsumer =
     vertex(matrix, x.toFloat() / 16f, y.toFloat() / 16f, z.toFloat() / 16f)
